requirements.txt
fastapi
pandas
numpy
scikit-surprise
gradio
datetime
mathplotlib